Extract editor output action helpers
