+++ CMakeLists.txt
@@ -160,6 +160,7 @@ target_sources(
         src/game_data_conditions_presentation.c
         src/game_data_controller_runtime.c
         src/game_data_controller_binding_helpers.c
+        src/game_data_editor_output_actions.c
         src/game_data_editor_player_start.c
         src/game_data_editor_selection_queries.c
         src/game_data_editor_debug_primitives.c
